refactor: 패키지 구성
