+++ src/main/java/christmas/model/menu/MenuType.java
@@ -1,4 +1,4 @@
-package christmas;
+package christmas.model.menu;
 
 public enum MenuType {
