+++ src/main/java/christmas/model/order/Orders.java
@@ -1,4 +1,4 @@
-package christmas;
+package christmas.model.order;
 
 import java.util.List;
