+++ src/main/java/christmas/view/InputValidator.java
@@ -1,4 +1,4 @@
-package christmas;
+package christmas.view;
 
 import java.util.Arrays;
