+++ src/main/java/christmas/view/InputView.java
@@ -1,4 +1,4 @@
-package christmas;
+package christmas.view;
 
 public class InputView {
