+++ src/test/java/christmas/model/OrderTest.java
@@ -1,5 +1,7 @@
-package christmas;
+package christmas.model;
 
+import christmas.model.menu.Menu;
+import christmas.model.order.Order;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
