+++ src/test/java/christmas/view/InputValidatorTest.java
@@ -1,4 +1,4 @@
-package christmas;
+package christmas.view;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
